rag_pipeline/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.0
tabula-py==2.8.0
numpy==1.24.0
openai==1.3.0
python-dotenv==1.0.0 